Add skipPOMs parameter to skip POM projects
Maven projects with packaging "pom" do not produce an executable artifact and
should be skipped.

This simplifies using the rpm-maven-plugin in multi-module builds in which the
configuration of the plugin is done in the parent project (usually just a POM
with the conventions for the sub-modules) and RPMs should be built only in the
sub-modules with packaging "jar" (or "war" or similar).