Add <dirmode> parameter to <mapping>
The 'dirmode' will be applied to directory SPEC entries that get added
implicitly while recursing into the indicated source directories.
If 'dirmode' is missing, 'filemode' will be used in an effort to
maintain backwards compatibility with the original behaviour.

When mapping directories recursively (recurseDirectories=true), we need
a separate set of permissions for files and directories. For many
packages, the user wants ordinary files without the executable bit,
while creating directories with the executable bit (so that they can be
 listed/enumerated)